Add deployment workflow (#34)
* Add deployment workflow

* Fix environment name casing

* Leverage latest version of FTP deploy action

The newer version of the original GitHub Action resolves a warning
regarding `known_hosts` configuration. The Automattic fork does not
appear to bring any meaningful value.

* Update deployment workflow naming

* Revert "Leverage latest version of FTP deploy action"

This reverts commit c84bbafebd2d940b1cefd3eb16ceba31d945a204.

* Add remote directory configuration

Sync the source files to the correct WordPress theme directory.

* Update deployment workflow remote directory

Attempt to fix missing file or directory error.

* Use absolute path for deployment workflow remote directory

Attempt to fix missing file or directory error.

* Place destination path in FTP server configuration

Remove usage of `--remote-root` to simplify configuration. Hoping this
results in the correct path on the server.

* Remove dry run deployment configuration

* Add temporary test file

Added to determine if the deployment workflow succeeded.

* Remove temporary test file

* Configure deployment workflow for main branch only